Refactors analytics modal to use separate confirmation modals that can be shared
