+++ src/enums/modal.ts
@@ -33,6 +33,7 @@ export enum ModalType {
   IncreaseAuthorizationSuccess = "INCREASE_AUTHORIZATION_SUCCESS",
   ConfirmDeauthorization = "CONFIRM_DEAUTHORIZATION",
   DeauthorizationCompleted = "DEAUTHORIZATION_COMPLETED",
-  DeauthorizationInitiated = "DeauthorizationInitiated",
+  DeauthorizationInitiated = "DEAUTHORIZATION_INITIATED",
   Analytics = "ANALYTICS",
+  FeedbackSubmission = "FEEDBACK_SUBMIT",
 }
