+++ src/types/modal.ts
@@ -45,6 +45,7 @@ import MapOperatorToStakingProviderModal from "../components/Modal/MapOperatorTo
 import MapOperatorToStakingProviderConfirmationModal from "../components/Modal/MapOperatorToStakingProviderConfirmationModal"
 import { MapOperatorToStakingProviderSuccess } from "../components/Modal/MapOperatorToStakingProviderSuccessModal"
 import AnalyticsModal from "../components/Modal/AnalyticsModal"
+import FeedbackSubmissionModal from "../components/Modal/FeedbackSubmissionModal"
 
 export const MODAL_TYPES: Record<ModalType, ElementType> = {
   [ModalType.SelectWallet]: SelectWalletModal,
@@ -88,6 +89,7 @@ export const MODAL_TYPES: Record<ModalType, ElementType> = {
   [ModalType.DeauthorizationCompleted]: DeauthorizationCompleted,
   [ModalType.DeauthorizationInitiated]: DeauthorizationInitiated,
   [ModalType.Analytics]: AnalyticsModal,
+  [ModalType.FeedbackSubmission]: FeedbackSubmissionModal,
 }
 
 export interface BaseModalProps {
